Move parachain components to separate folder
