+++ cumulus/polkadot-parachain/src/rpc.rs
@@ -18,9 +18,9 @@
 
 #![warn(missing_docs)]
 
-use crate::{
-	common::ConstructNodeRuntimeApi,
-	service::{ParachainBackend, ParachainClient},
+use crate::common::{
+	parachain::{ParachainBackend, ParachainClient},
+	BuildRpcExtensions, ConstructNodeRuntimeApi,
 };
 use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApiServer};
 use parachains_common::{AccountId, Balance, Block, Nonce};
@@ -35,15 +35,6 @@ use substrate_state_trie_migration_rpc::{StateMigration, StateMigrationApiServer
 /// A type representing all RPC extensions.
 pub type RpcExtension = jsonrpsee::RpcModule<()>;
 
-pub(crate) trait BuildRpcExtensions<Client, Backend, Pool> {
-	fn build_rpc_extensions(
-		deny_unsafe: DenyUnsafe,
-		client: Arc<Client>,
-		backend: Arc<Backend>,
-		pool: Arc<Pool>,
-	) -> sc_service::error::Result<RpcExtension>;
-}
-
 pub(crate) struct BuildEmptyRpcExtensions<RuntimeApi>(PhantomData<RuntimeApi>);
 
 impl<RuntimeApi>
@@ -53,7 +44,7 @@ impl<RuntimeApi>
 		sc_transaction_pool::FullPool<Block, ParachainClient<RuntimeApi>>,
 	> for BuildEmptyRpcExtensions<RuntimeApi>
 where
-	RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<RuntimeApi>> + Send + Sync + 'static,
+	RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<RuntimeApi>>,
 {
 	fn build_rpc_extensions(
 		_deny_unsafe: DenyUnsafe,
@@ -74,7 +65,7 @@ impl<RuntimeApi>
 		sc_transaction_pool::FullPool<Block, ParachainClient<RuntimeApi>>,
 	> for BuildParachainRpcExtensions<RuntimeApi>
 where
-	RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<RuntimeApi>> + Send + Sync + 'static,
+	RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<RuntimeApi>>,
 	RuntimeApi::RuntimeApi: pallet_transaction_payment_rpc::TransactionPaymentRuntimeApi<Block, Balance>
 		+ substrate_frame_rpc_system::AccountNonceApi<Block, AccountId, Nonce>,
 {
